Add 'omnisharp.autoStart' option that allows users to control whether OmniSharp is started automatically or not
