+++ package.json
@@ -99,6 +99,11 @@
             "verbose"
           ],
           "description": "Specifies the level of logging output from the OmniSharp server."
+        },
+        "omnisharp.autoStart": {
+          "type": "boolean",
+          "default": true,
+          "description": "Specifies whether the OmniSharp server will be automatically started or not. If false, OmniSharp can be started with the 'Restart OmniSharp' command"
         }
       }
     },
