+++ src/omnisharp/omnisharp.ts
@@ -13,12 +13,6 @@
 import * as fs from 'fs-extra-promise';
 import * as path from 'path';
 
-export interface Options {
-    path?: string;
-    useMono?: boolean;
-    loggingLevel?: string;
-}
-
 export enum Flavor {
     CoreCLR,
     Mono,
